Created models.py and added required imports
