+++ mde/models.py
@@ -1,0 +1,8 @@
+from mde import db
+
+class User(db.Model):
+    id = db.Column(db.Integer(), primary_key=True)
+    name = db.Column(db.String(length=30), nullable=False, unique=True)
+    email_address = db.Column(db.String(length=50), nullable=False, unique=True)
+    password_hash = db.Column(db.String(length=60), nullable=False)
+    parent_email_address = db.Column(db.String(length=50), nullable=True, unique=True)
